+++ src/main/java/com/ht/project/snsproject/model/Pagination.java
@@ -1,10 +1,8 @@
 package com.ht.project.snsproject.model;
 
-import lombok.Builder;
 import lombok.Value;
 
 @Value
-@Builder
 public class Pagination {
 
   Integer cursor;
@@ -13,7 +11,6 @@ public class Pagination {
 
   public static Pagination pageInfo(Integer cursor) {
 
-    return Pagination.builder()
-            .cursor(cursor).build();
+    return new Pagination(cursor);
   }
 }
