--- src/main/java/com/ht/project/snsproject/model/good/GoodPushedStatusesParam.java
@@ -1,15 +1,0 @@
-package com.ht.project.snsproject.model.good;
-
-import lombok.Builder;
-import lombok.Value;
-
-import java.util.List;
-
-@Value
-@Builder
-public class GoodPushedStatusesParam {
-
-  List<Integer> feedIds;
-
-  String userId;
-}
