+++ src/main/resources/application-default.properties
@@ -1,12 +1,22 @@
 
-spring.datasource.hikari.jdbc-url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
-spring.datasource.hikari.username=root
-spring.datasource.hikari.password=rootqwer
+spring.datasource.master.hikari.jdbc-url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
+spring.datasource.master.hikari.username=root
+spring.datasource.master.hikari.password=rootqwer
 
-spring.datasource.url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
-spring.datasource.username=root
-spring.datasource.password=rootqwer
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.master.url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
+spring.datasource.master.username=root
+spring.datasource.master.password=rootqwer
+spring.datasource.master.driver-class-name=com.mysql.cj.jdbc.Driver
+
+
+spring.datasource.slave.hikari.jdbc-url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
+spring.datasource.slave.hikari.username=root
+spring.datasource.slave.hikari.password=rootqwer
+
+spring.datasource.slave.url=jdbc:mysql://localhost:3306/flab?useUnicode=true&serverTimezone=GMT&autoReconnect=true
+spring.datasource.slave.username=root
+spring.datasource.slave.password=rootqwer
+spring.datasource.slave.driver-class-name=com.mysql.cj.jdbc.Driver
 
 redis.session.host=localhost
 redis.session.password=
